Update test parameters and error handling in CarSearchService
The CarSearchService's 'search2' method in test now uses a different search parameter. Moreover, an additional check has been added in 'carDetails' method of CarSearchService for cases where the 'Vin' key is present. Following this, an error message is returned when the response size is less than or equal to 1. Modifications were also made to ArkService class to improve error handling when no car is found for provided registration number.