Refactor RabbitMQ messaging to use Spring AMQP
Changed the messaging infrastructure from using spring-cloud-stream to Spring AMQP for better handling of RabbitMQ operations. The RabbitMQ publisher and consumer classes have been updated accordingly. This shift necessitated the removal of certain lines of code from the application.yml file, as well as the addition of a new RabbitMQConfig file for managing associated configurations. The adjustments in the IntegrationTest and pom.xml files reflect these updates.