Improve automatic fetching functionality in ElectricityBotService
The changes apply to the creation of the inline keyboard message in 'ElectricityBotService.java'. Now, the message can vary depending on if automatic fetching is enabled. If so, the detected plate number is directly shown to the user. Otherwise, a prompt asking if the user would like to check the potential plate number is displayed. The automatic fetching activation check was added in the SendMessage creation process.