Add Automaks functionality and update CacheService
Added Automaks function and corresponding buttons in the ElectricityBotService. Enable/disable mechanism is now stored in the CacheService and can be queried through new methods. The implementation of the Automaks feature will offer a new approach for anonymized data sharing while enhancing user interaction options. CacheService is improved to handle the new Automaks status as well.