Refactor sendMessage method and handle messageId
Modified the sendMessage function in the ElectricityBotService to return the sent Message and adjusted the way messageIds are handled. Enhanced the service to now return the message sent, allowing greater control over the Message object. Changes also cover better handling of the messageId for registration plates which provides more flexibility in data manipulation and management.