Refactor CarSearchService and update RedisConfig
Refactored CarSearchService to add an update listener and modify future operations to enhance performace. The changes also include the implementation of the scrapeNinjaDetails feature and changing the return response structure handling. Extended RedisConfig cache naming to improve clarity and avoid confusion. Overall, these changes aim to improve the performance, readability and maintainability of the system.