Update time duration threshold and variable name in ElectricityBotService
The time duration threshold is now decreased from 15 seconds to 10 in ElectricityBotService. The variable 'aDouble' which was used to store duration in seconds has been renamed to 'animationDuration'. This improves clarity by making the threshold and variable name more reflective of their functional role in the system. This should ease understanding and future maintenance.