Modify Euribor rate check and message conditions
Changes have been made to the SchedulingService. The "checkAndSendEuriborRate" method has been updated to change the schedule from running every 60 minutes to running on the 7th minute of every hour. Furthermore, a new condition has been added to check if an Euribor message can be sent today before proceeding with the process. If not, it logs that no new Euribor rate was detected and does not send a message.