Modify Euribor rate fetching and messaging logic
The commit updates the "checkAndSendEuriborRate" method in the SchedulingService. Previously, the method was set to run every 60 minutes and fetch the latest Euribor rate while updating the cache, and then it would send an alert if the rate has changed even if the daily messaging limit was reached. Now, it's scheduled to run on the 15th minute of every hour, checks the daily message limit before fetching the rate or sending a message, and no longer updates the cache while fetching the latest rate.