Modify SchedulingService to use ElectricityBotService
The SchedulingService class has been updated to include the ElectricityBotService. In addition to adding ElectricityBotService as a resource, the code now uses it to send a message with the current Euribor rate. The cron schedule for checking and sending the Euribor rate was also adjusted to a new timing. This update facilitates more efficient and centralised message sending.