Add online check to application and revise timeout constant
Added an online check function to the application, which determines if MacBook is online. Created new RabbitMQ queues and methods for supporting this function. Also, cleaned up the TIMEOUT constant in the ChatService class by removing unnecessary comment about its value being in milliseconds. This expands the functionality and improves the readability of the code.