Remove car search key and update timing of listener update
Code in CarSearchService was adjusted to update the listener immediately after gathering car prices, eliminating redundancy. Additionally, the carPriceFuture was removed from the CompletableFuture.allOf() method as its results are already used prior to this point. This rework enhances the efficiency and responsiveness of the function. Sensitive key values have also been removed to improve security.