Improve PredictService with duration logging
An enhancement has been made to the PredictService by introducing logging that keeps track of the duration of each prediction. This comes as a useful tool for tracking performance and tracing issues in case the prediction process takes unusually long. TimeUtility class is also imported to aid in the computation of prediction duration.