Update Docker commands and container names in CI workflow
The continuous integration workflow has been updated to correctly remove the `elektrihind_main` and `captcha_solver_main` Docker containers. In the meantime, an unnecessary command to remove the `elektrihind-elektrihind-app-1` container was also deleted from the script. The aim of these changes is to improve the cleanliness of the test environment and ensure the correct images are being used during testing.