Refactor and expand Apollo Kino services
Implemented an improved system for managing user sessions and booking cinema tickets in Apollo Kino. This system now includes session expiration handling as well as repeated booking attempts if initial attempts are unsuccessful. A key change is the switch from UUIDs to Integer for session IDs. Refactored several configuration files and existing classes to integrate with the new changes and created new classes for the re-booking and cinema services.