Renamed method and added limit to active bookings
The method 'getSessionCount' has been renamed to 'getActiveBookingCount' to better reflect its content. Additionally, a condition has been added to the APOLLO_KINO case in 'ElectricityBotService.java' to restrict the number of active bookings to a maximum of 4 per user. Now, if there are more than four bookings, the bot sends a message to the user requesting them to try again later or cancel existing bookings.