Correct 'applo-kino' typo and restrict Telegram bot usage
This commit fixes the typo 'applo-kino' to 'apollo-kino' across various properties files and code that reference these properties. This change is also reflected in environment variable configuration in the Docker setup and GitHub Actions. Additionally, restrictions are added on who can use the Telegram bot service to a predefined list of usernames.