Add Automaks service to calculate vehicle tax
Integrate Automaks service for tax calculation and refactor ArkService to use it. Also, refactor the AutomaksClient for this change. The changes also include modifying the docker configurations and updating the test resources to support the new service integration. Finally, some minor modifications have been made in related classes such as ReBookingService and EasyOcrClient.