Improve ApolloKinoService session handling and rebooking
Enhancements were made to ApolloKinoService's session handling to optimize certain actions, and an "optional reBook" feature was added for a more robust booking system. The system now effectively checks for terms acceptance before logging in and the function for rebooking sessions was optimized in the "ReBookingService". Also, the server's port in 'application.yml' is now customizable.